test: merge the service timeout suites into one asset
Why: the favorites timeout tests duplicated the reverse ones. The two
slow web service mocks differed by a single column, and each test class
launches its own container stack, so the copy bought nothing.

One `service_timeout` asset and one test class now cover the timeout
behaviour of both services, which removes a stack launch and gives the
lookup service an obvious home when its coverage is added.
