--- integration_tests/assets/docker-compose.reverse_timeout.override.yml
@@ -1,19 +1,0 @@
-services:
-  sync:
-    depends_on:
-      - auth
-      - dird
-      - db
-      - ws
-    environment:
-      TARGETS: "dird:9489 auth:9497 db:5432 ws:9485"
-  ws:
-    image: wazoplatform/wazo-dird-ws-mock:local
-    ports:
-      - "9485"
-    volumes:
-      - "./scripts/asset.reverse_timeout.slow_ws.py:/tmp/ws.py"
-    command: "python3 /tmp/ws.py"
-  dird:
-    volumes:
-      - "./tmp/data/asset.reverse_timeout.fast.csv:/tmp/data/fast.csv"
