--- integration_tests/assets/tmp/data/asset.reverse_timeout.fast.csv
@@ -1,2 +1,0 @@
-number,firstname,lastname
-5559999999,Bob,Fast
